multi-agent/backend/requirements.txt
# Web framework
fastapi>=0.115,<0.116
uvicorn[standard]>=0.32,<0.33
sse-starlette>=2.1,<3
pydantic>=2.10,<3

# Agents — let pip pick a compatible langchain-core transitively.
langgraph>=0.2.55,<0.3
langchain>=0.3.10,<0.4
langchain-groq>=0.2.1,<0.3
langchain-mcp-adapters>=0.0.5,<0.1
langchain-tavily>=0.2.0,<0.3

# Misc
python-dotenv>=1.0,<2
httpx>=0.28,<0.29

# Tests
pytest>=8.3,<9
pytest-asyncio>=0.24,<0.25